+++ R/artifact_edition.R
@@ -36,7 +36,7 @@
 #'
 #' # Signals with artifacts are turned into NA values:
 #' faces_clean <- faces_seg_artif %>%
-#'   eeg_events_to_NA(.type == "artifact", .entire_seg = TRUE, .all_chs = FALSE, .drop_events = TRUE)
+#'   eeg_events_to_NA(.type == "artifact", .entire_seg = TRUE, .drop_events = TRUE)
 #' }
 #'
 NULL
@@ -250,9 +250,10 @@ eeg_artif_peak.eeg_lst <- function(.data,
 #'
 #' @param x An `eeg_lst` object.
 #' @param ... Description of the problematic event.
-#' @param .all_chs If set to `TRUE`, it will consider samples from all channels (Default:  `.all_chs = FALSE`).
+#' @param .n_chs  If set to `NULL` (default), it will only set `NA` to the relevant channel.If set to a number `N`, it will set all the channels of the entire segment or interval to `NA`, if `N` or more channels have a certain event, and otherwise only the relevant channel.
 #' @param .entire_seg If set to `FALSE`, it will consider only the marked part of the segment, otherwise it will consider the entire segment (Default: .entire_seg = TRUE).
 #' @param .drop_events If set to `TRUE` (default), the events that were used for setting signals to NA, will be removed from the events table.
+#' @param .all_chs Deprecated.
 #'
 #' @family events functions
 #'
@@ -261,12 +262,12 @@ eeg_artif_peak.eeg_lst <- function(.data,
 #'
 #' # Signals with artifacts are turned into NA values:
 #' faces_clean <- faces_seg_artif %>%
-#'   eeg_events_to_NA(.type == "artifact", .entire_seg = TRUE, .all_chs = FALSE, .drop_events = TRUE)
+#'   eeg_events_to_NA(.type == "artifact", .entire_seg = TRUE, .drop_events = TRUE)
 #'
 #'
 #' # Specific segments are turned into NA values:
 #' faces_clean <- faces_seg_artif %>%
-#'   eeg_events_to_NA(.id %in% c(1:3), .entire_seg = TRUE, .all_chs = FALSE, .drop_events = TRUE)
+#'   eeg_events_to_NA(.id %in% c(1:3), .entire_seg = TRUE, .drop_events = TRUE)
 #' }
 #'
 #' @return An eeg_lst.
@@ -274,9 +275,10 @@ eeg_artif_peak.eeg_lst <- function(.data,
 eeg_events_to_NA <-
   function(x,
            ...,
-           .all_chs = FALSE,
+           .n_chs = NULL,
            .entire_seg = TRUE,
-           .drop_events = TRUE) {
+           .drop_events = TRUE,
+           .all_chs = FALSE) {
     UseMethod("eeg_events_to_NA")
   }
 
@@ -284,20 +286,34 @@ eeg_events_to_NA <-
 eeg_events_to_NA.eeg_lst <-
   function(x,
            ...,
-           .all_chs = FALSE,
+           .n_chs = NULL,
            .entire_seg = TRUE,
-           .drop_events = TRUE) {
+           .drop_events = TRUE,
+           .all_chs = FALSE) {
     dots <- rlang::enquos(...)
 
     signal <- data.table::copy(x$.signal)
     # dots <- rlang::quos(.type == "Bad Interval")
+    #dots <- rlang::quos(.type == "Bad")
 
+    # 
     # Hack for match 2 columns with 2 columns, similar to semi_join but allowing
     # for assignment
     baddies <- filter.(x$.events, !!!dots)
-
-    if (.all_chs) {
-      baddies <- mutate.(baddies, .channel = NA_character_)
+    if (!missing(".all_chs")){
+      warning("The argument `.all_chs` is deprecated.\n",
+              "Set `.n_chs = NULL` to reproduce the behavior of `.all_chs = FALSE`. ",
+              "Set `.n_chs = 1` to reproduce the behavior of `.all_chs = TRUE`.")
+      if(.all_chs) .n_chs <- 1 else .n_chs <- NULL
+    }
+    if (!is.null(.n_chs)) {
+    ## this doesn't work
+            # baddies <- baddies %>% mutate.(.channel = 
+            #                 tidytable::case_when.(n() > .n_chs ~ NA_character_,
+            #                                       TRUE ~ .channel),
+            #                                       .by = ".id")
+      baddies[baddies[, .I[.N >= .n_chs], .id]$V1, .channel := NA_character_] 
+            # baddies <- mutate.(baddies, .channel = NA_character_)
     }
 
     # For the replacement in parts of the segments