+++ NEWS.md
@@ -3,6 +3,7 @@
 - Signal filtering is slightly faster.
 - BUGS:
   - some brainvision files produced by eeglab couldn't be read.
+- The argument `.all_chs` of `eeg_events_to_NA` was deprecated in favor of the more flexible `.n_chs`.
 
 # eeguana 0.1.6.9000